feat(hv): add hypervisor support for GICv2
- Add GicHypervisorInterface struct and related functionality
- Implement methods for accessing hypervisor-specific GIC registers
- Add support for virtual interrupt management and maintenance interrupts
- Update Cargo.toml to include 'hv' feature
- Refactor gic_v2 module to separate hypervisor-related code
