+++ src/regs/gich/mod.rs
@@ -1,0 +1,18 @@
+mod gich_apr;
+mod gich_eisr;
+mod gich_elrsr;
+mod gich_hcr;
+mod gich_lr;
+mod gich_misr;
+mod gich_vmcr;
+mod gich_vtr;
+
+// Export the common interfaces of all modules.
+pub use gich_apr::*;
+pub use gich_eisr::*;
+pub use gich_elrsr::*;
+pub use gich_hcr::*;
+pub use gich_lr::*;
+pub use gich_misr::*;
+pub use gich_vmcr::*;
+pub use gich_vtr::*;
