+++ src/regs/mod.rs
@@ -1,3 +1,6 @@
 mod gicd_sgir;
 
 pub use gicd_sgir::*;
+
+#[cfg(feature = "hv")]
+pub mod gich;
